backend/requirements.txt
asgiref==3.2.3
Django==3.0.3
pytz==2019.3
sqlparse==0.3.0
uWSGI==2.0.18
django-filter==2.2.0
djangorestframework==3.11.0
markdown==3.2.1
black==19.10b0
